requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
opencv-python
python-multipart
requests
python-dotenv
inference-sdk
jinja2